+++ Engine/Vec2.h
@@ -4,18 +4,18 @@ class Vec2
 {
 public:
 	Vec2() = default;
-	Vec2( float x_in,float y_in );
+	Vec2( double x_in,double y_in );
 	Vec2 operator+( const Vec2& rhs ) const;
 	Vec2& operator+=( const Vec2& rhs );
-	Vec2 operator*( float rhs ) const;
-	Vec2& operator*=( float rhs );
+	Vec2 operator*( double rhs ) const;
+	Vec2& operator*=( double rhs );
 	Vec2 operator-( const Vec2& rhs ) const;
 	Vec2& operator-=( const Vec2& rhs );
-	float GetLength() const;
-	float GetLengthSq() const;
+	double GetLength() const;
+	double GetLengthSq() const;
 	Vec2& Normalize();
 	Vec2 GetNormalized() const;
 public:
-	float x;
-	float y;
+	double x;
+	double y;
 };
